move LIP to protocol
